VT6130 datasheet was wrong. If VT6130 receive a jumbo frame the controller will split the jumbo frame into multiple RX buffers. However it seems the hardware always dma the frame to 8 bytes boundary for the split frames. Only the first part of the fragment can have 4 byte alignment and subsequent buffers should be 8 bytes aligned. Change RX buffer the alignment requirement to 8 bytes from 4 bytes.
Former-commit-id: 0d7faed4a69b66c50628c0e69e76d9577b1ec31c

+++ sys/dev/vge/if_vgevar.h
@@ -53,7 +53,7 @@ begin_define
 define|#
 directive|define
 name|VGE_RX_BUF_ALIGN
-value|sizeof(uint32_t)
+value|sizeof(uint64_t)
 end_define
 
 begin_comment
